Adds initial file structure refactor.
