+++ setup.py
@@ -1,12 +1,13 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 setup(
     name="measuresoftgram",
     version="1.0.0",
     extras_require={"dev": ["pytest", "pytest-cov", "setuptools", "wheel"]},
     install_requires=["inquirer", "requests"],
+    packages=find_packages(),
     # To provide executable scripts, use entry points in preference to the
     # "scripts" keyword. Entry points provide cross-platform support and allow
     # pip to create the appropriate form of executable for the target platform.
-    entry_points={"console_scripts": ["measuresoftgram=measuresoftgram.cli:main"]},
+    entry_points={"console_scripts": ["measuresoftgram=src.cli.cliRunner:main"]},
 )
