+++ src/cli/exceptions.py
@@ -1,0 +1,4 @@
+class MeasureSoftGramCLIException(Exception):
+    ''' Base MeasureSoftGram CLI Exception '''
+    pass
+
